requirements.txt
django
gunicorn
whitenoise
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
django-cors-headers
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
